refactor(edge-functions): move CORS logic into _shared/cors.ts
Keeps auth.ts scoped to authentication/admin-client concerns.
